+++ tests/Tests/PHPUnit/AbstractTestCasesTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\AvtoDev\DevTools\Tests\PHPUnit;
 
+use AvtoDev\DevTools\Tests\PHPUnit\Traits\LaravelLogFilesAssertsTrait;
 use Tests\AvtoDev\DevTools\AbstractTestCase;
 use AvtoDev\DevTools\Tests\PHPUnit\Traits\CreatesApplicationTrait;
 use AvtoDev\DevTools\Tests\PHPUnit\Traits\InstancesAccessorsTrait;
@@ -36,6 +37,7 @@ class AbstractTestCasesTest extends AbstractTestCase
     public function testAbstractLaravelTestCase()
     {
         $instance = new class extends \AvtoDev\DevTools\Tests\PHPUnit\AbstractLaravelTestCase {
+            use CreatesApplicationTrait;
         };
 
         $this->assertInstanceOf(\PHPUnit\Framework\TestCase::class, $instance);
@@ -46,6 +48,7 @@ class AbstractTestCasesTest extends AbstractTestCase
             InstancesAccessorsTrait::class,
             CreatesApplicationTrait::class,
             LaravelEventsAssertionsTrait::class,
+            LaravelLogFilesAssertsTrait::class,
         ]);
     }
 }