+++ CHANGELOG.md
@@ -1,23 +1,34 @@
 # Changelog
 
+## v1.1.4
+
+### Added
+
+- Additional method `getApplicationBootstrapFiles()` into `CreatesApplicationTrait`
+- Trait `LaravelLogFilesAssertsTrait`
+
 ## v1.1.3
 
 ### Fixed
+
 - Fixed asserts `assertEmptyString`, `assertNotEmptyString`, `assertClassExists` - now works correctly with arrays of input values
 
 ## v1.1.2
 
 ### Fixed
+
 - Fixed application bootstrap file path for using in packages in `CreatesApplicationTrait`
 
 ## v1.1.1
 
 ### Fixed
+
 - Fixed exceptions handler in `AbstractTestsBootstrapper`
 
 ## v1.1.0
 
 ### Added
+
 - Assert method `assertIsInteger`
 
 ## v1.0.0